Fix merge conflicts

+++ app/Extensions/Backups/Adapter/S3BackupAdapter.php
@@ -112,7 +112,7 @@ class S3BackupAdapter implements BackupAdapter
      */
     private function getConfiguredMaxPartSize(): int
     {
-        $maxPartSize = (int) config('backups.max_part_size', self::DEFAULT_MAX_PART_SIZE);
+        $maxPartSize = config('backups.max_part_size', self::DEFAULT_MAX_PART_SIZE);
         if ($maxPartSize <= 0) {
             $maxPartSize = self::DEFAULT_MAX_PART_SIZE;
         }
